+++ HISTORY.rst
@@ -6,6 +6,10 @@ Release History
 Unreleased
 ++++++++++
 
+**Data Migration**
+
+* ALCN-1109: Create partial receptions on imported in progress purchases
+
 **Features and Improvements**
 
 **Bugfixes**